feat(gc): add prune-only worker subcommand for generation links
Extract the "prune generation links across all registered managed
environments" logic out of `flox gc` into a shared
`prune_registered_environments`, and add a hidden `flox
prune-generation-links` worker that runs only that prune — NOT the nix
store GC (flox#4332 F3, PR3b).

The worker takes a non-blocking lock (`generation-prune.lock` in the
runtime dir) and exits if another prune holds it, so the activation
executives that will fire it periodically don't race to remove the same
links. `flox gc` keeps calling the shared prune directly before its store
GC.
